==> 02_STOPPD_xnat.Rmd <==
---
title: "02_STOPPD_xnat"
author: "Navona"
written: January 18, 2018  
last run: Sys.Date()
output: html_document
---

#This script pulls in and cleans up the naming of STOPPD scans as they exist in XNAT. At earlier stages, this script helped us identify naming errors in XNAT (all have since been fixed).

#Purpose: The contents of XNAT will, in other scripts, be checked against (1) the scans we have in our file system, to ensure that there are no discrepancies between these databases, and also against (2) our subject inclusion list.

```{r setup}
library('stringi')
library('stringr')
library('plyr')
library('tidyr')
```

```{r import_data}

#import spreadsheets (exported from XNAT)
xnat_camh <- read.csv('../data/xnat_cmh_2018-01-25.csv')
xnat_nki <- read.csv('../data/xnat_nki_2018-01-25.csv')
xnat_pitt <- read.csv('../data/xnat_pmc_2018-01-25.csv')
xnat_umass <- read.csv('../data/xnat_umas_2018-01-25.csv')

#combine XNAT spreadsheets, take only ID and date columns
xnat <- Reduce(function(x, y) merge(x, y, all=TRUE), list(xnat_camh, xnat_nki, xnat_pitt, xnat_umass))
xnat <- xnat[c('MR.ID', 'Date') ]

#cleanup
rm (xnat_camh, xnat_nki, xnat_pitt, xnat_umass)

#import spreadsheet of data in file system (made in script 01_STOPPD_terminal)
terminal <- read.csv('../generated_csvs/terminal_clean_2018-01-25.csv')

```

```{r data_cleaning}

#remove all CAMH scans with '00' as timepoint (NOTE: '00' this is a consequence of creative naming to account for MRS scans)
xnat$timepoint <- str_sub(xnat$MR.ID, start= -2) #make column with timepoint data
xnat <- xnat[-grep('00', xnat$timepoint),] #remove those with 00

#cut out timepoint info from subject ID string - now meaningless - and remove timepoint column
xnat$MR.ID  <- str_sub(xnat$MR.ID, 1, -4)
xnat <- xnat[, -grep('timepoint', colnames(xnat))]

#cut out study and site info from subject ID string - not needed
xnat$MR.ID <- substring(xnat$MR.ID, 12)

#make a new column for session component of ID
xnat$session <- str_sub(xnat$MR.ID, -2)
table(xnat$session)

#cut out session from subject ID string - not needed
xnat$MR.ID <- str_sub(xnat$MR.ID, 1, -4)

#make a new column that captures alphabetic component of ID ('R')
xnat$contains_R <- grepl('R', xnat$MR.ID, fixed=TRUE) 

#cut out the 'R' in some participant IDs (indicates repeat for controls)
xnat$MR.ID <- gsub("[R]", "", xnat$MR.ID)

#make a variable that combines unique ID and session number
xnat$id_session <- paste(xnat$MR.ID, '_', xnat$session, sep='')

#check for consistency between file system and XNAT 
X <- terminal$id_session %in% xnat$id_session 
  which(X == FALSE) #identical

Y <- xnat$id_session %in% terminal$id_session 
  which(Y == FALSE) #identical

#write csv
write.csv(xnat, '../generated_csvs/xnat_clean_2018-01-25.csv', row.names=FALSE)

#cleanup
rm(terminal, xnat)
```
