==> 06a_STOPPD_CT-left_2018-09-09.Rmd <==
---
title: |
  | <br><br><br><br>STOPPD cortical thickness analysis
  | (left hemisphere only)
written: 2018-03-08
edited: 2018-09-09
lastrun: Sys.Date()
output: 
  pdf_document:
    latex_engine: xelatex
    toc: yes
    toc_depth: '2'

---

#This script analyses cortical thickness in the left hemisphere. 

```{r set_up, include = FALSE}

#load libraries
library(tidyverse)
library(lme4)
library(lmerTest)
library(growthmodels)
library(reshape2)
library(ggplot2)

#bring in data 
df <- read.csv('../generated_csvs/STOPPD_errorCases_2018-02-09.csv') #generated by 05_STOPPD_error in prepartion of Jason Lerch meeting

```

```{r datacleaning, include=FALSE}

#review of missing data
  which(is.na(df[,-6]), arr.ind=TRUE) #ignoring col. 6 (off label), identify location of NAs - we have no missing CT

#make sure that dateDiff is a number, not an interger
  df$dateDiff <- as.numeric(df$dateDiff)

#make a difference score 
  df$LDifference <- df$LThickness.02 - df$LThickness.01 #we're only concerned with left thickness (as per Nick's request)

#restructure data for RCT completers' only (N=41)
  RCT_CT <- df[(df$category == "RCT"),] 

#restructure data for RCT & Relapse participants (N=74)
  RCTRelapse_CT <- df %>%
    gather(thick_oldcolname, thickness, LThickness.01, LThickness.02) %>%
    mutate(model_days = if_else(thick_oldcolname == "LThickness.01", 1, dateDiff))

#write out clean dataframe
  write.csv(RCT_CT, '../generated_data/df_leftCT.csv', row.names=FALSE)

```

## RCT only

```{r RCT_CT}

#boxplot of difference in thickness (y axis) by randomization group (x axis)
  (plot <- ggplot(RCT_CT, aes(x= randomization, y = LDifference)) + 
     geom_boxplot(outlier.shape = NA) + geom_jitter() +
     ggtitle("Cortical thickness (left hemisphere)") +
     xlab("randomization condition") +
     ylab("difference between timepoints")) 

#run linear model without covariates
  fit_rct <- lm(LDifference ~ randomization + (1|STUDYID), data= RCT_CT)
  print(fit_rct)
  summary(fit_rct)

#run linear model with covariates of sex and age
  fit_rct <- lm(LDifference ~ randomization + sex + age + (1|STUDYID), data= RCT_CT)
  print(fit_rct)
  summary(fit_rct)

```

## RCT only, exclude outlier

```{r RCT_CT_exclude outlier}

#exclude outlier (participant 210030)
 RCT_CT_no210030 <- RCT_CT[!(RCT_CT$STUDYID == "210030"),]

#boxplot of difference in thickness (y axis) by randomization group (x axis)
  (plot <- ggplot(RCT_CT_no210030, aes(x= randomization, y = LDifference)) +
           geom_boxplot(outlier.shape = NA) + geom_jitter() +
           ggtitle("Cortical thickness (left hemisphere), outlier removed") +
           xlab("randomization condition") +  
           ylab("difference between timepoints")) 

#run linear model without covariates
  fit_rct_no210030 <- lm(LDifference ~ randomization + (1|STUDYID), data= RCT_CT_no210030)
  print(fit_rct_no210030)
  summary(fit_rct_no210030)

#run linear model with covariates of sex and age
  fit_rct_no210030 <- lm(LDifference ~ randomization + sex + age + (1|STUDYID), data= RCT_CT_no210030)
  print(fit_rct_no210030)
  summary(fit_rct_no210030)

```

## RCT & Relapse (with time as factor)

```{r RCTRelapse_CT, warning = FALSE}

#plot all data, including outlier (participant 210030)
  RCTRelapse_CT %>%
   ggplot(aes(x=model_days, y=thickness, colour=randomization)) + 
   geom_point() + 
   geom_line(aes(group=STUDYID), alpha = 0.5) + 
   geom_smooth(method="lm", formula=y~poly(x,1)) +
   ggtitle("Cortical thickness in left hemisphere over time") +
   xlab("days between MRIs") +  
   ylab("thickness value") +
   theme_minimal()
  
#run mixed linear model, with covariates
  fit_all <- lmer(thickness ~ randomization*model_days + sex + age + (1|STUDYID), data= RCTRelapse_CT)
  summary(fit_all)

```

## RCT & Relapse (with time as factor), exclude outlier

```{r RCTRelapse_CT_exclude outlier, warning = FALSE}

#do same analysis as above with outlier removed (participant 210030)
  RCTRelapse_CT_no210030 <- RCTRelapse_CT[!(RCTRelapse_CT$STUDYID == "210030"),]

#plot data excluding outlier
  RCTRelapse_CT_no210030 %>%
   ggplot(aes(x=model_days, y=thickness, colour=randomization)) + 
   geom_point() + 
   geom_line(aes(group=STUDYID), alpha = 0.5) + 
   geom_smooth(method="lm", formula=y~poly(x,1)) +
   ggtitle("Cortical thickness in left hemisphere over time, outlier removed") +
   xlab("days between MRIs") +  
   ylab("thickness value") +
   theme_minimal()

#run mixed linear model, with covariates, as above
  fit_all_no210030 <- lmer(thickness ~ randomization*model_days + sex + age + (1|STUDYID), data= RCTRelapse_CT_no210030)
  summary(fit_all_no210030)

#cleanup
  rm(df, fit_all, fit_all_no210030, fit_rct, fit_rct_no210030, plot, RCT_CT, RCT_CT_no210030, RCTRelapse_CT, RCTRelapse_CT_no210030)
  
```
