==> 06b_STOPPD_MD-lEC_2018-09-09.Rmd <==
---
title: |
  | <br><br><br><br>STOPPD mean diffusivity analysis
  | (left entorhinal cortex)
written: 2018-03-08
edited: 2018-09-09
lastrun: Sys.Date()
output: 
  pdf_document:
    latex_engine: xelatex
    toc: yes
    toc_depth: '2'
---

#This script analyses mean diffusivity in the left entorhinal cortex. 

```{r set_up, include = FALSE}

#load libraries
library(tidyverse)
library(lme4)
library(lmerTest)
library(growthmodels)
library(reshape2)
library(ggplot2)
library(stringi)

#bring in MD data (from the filesystem)
MD <- read.csv('../data/enigmaDTI-MD_2018-09-05.csv')

#bring in subject info (generated by 03_STOPPD_masterDF.Rmd)
df <- read.csv('../generated_csvs/STOPPD_masterDF_2018-09-06.csv')

```

```{r MD_datacleaning_general, include = FALSE}

#pull timepoint info from subject ID
MD$timepoint <- str_sub(MD$id, -2, -1)

#remove scans from unwanted timepoints
MD <- MD[!(MD$timepoint == '03'),]
MD <- MD[!(MD$timepoint == '00'),]

#clean up subject ID
MD$id <- substring(MD$id, 12) #remove first 12 chars
MD$id <- str_sub(MD$id, 1, -4) #remove last 3 chars

#move from wide to long format
MD <- reshape(MD, idvar = "id", timevar = "timepoint", direction = "wide")

#limit data to just those who completed both first and second scan
completers <- df$STUDYID[df$second_complete == 'Yes'] #make list of complters
MD <- MD[ MD$id %in% completers, ] #keep only names on that list

#review of missing data
which(is.na(MD), arr.ind=TRUE) #identify location of NAs 
  #310051 missing EC_MD.01, reason: ??? COME BACK TO THIS - Erin re-running
  #220009 missing EC_MD.01, reason: scan incomplete (participant refused)
  #410012 missing EC_MD.02, reason: scan incomplete (system failure; also had meningioma)

#merge subject info
df <- merge(MD, df[, c("STUDYID", 'sex', 'age', 'randomization', 'second_timepoint', 'first_date', 'second_date')], by.x = 'id', by.y = 'STUDYID')

#rename timepoint variable for clarity
colnames(df)[colnames(df)=="second_timepoint"] <- "category" 

#make a datediff column for time between scans
df$dateDiff <- as.numeric(round(difftime(df$second_date, df$first_date, units = "days"), 0))

#write out clean MD speadsheet (required for subsequent MD analyses)
write.csv(df, '../generated_csvs/STOPPD_MDclean.csv', row.names = FALSE)

```

```{r datacleaning_lEC, include = FALSE}

#take only the metrics we care about
df <- df[, c("id", 'sex', 'age', 'randomization', 'category', 'dateDiff', 'EC.L_MD.01', 'EC.L_MD.02')]

#make sure that the ID is numeric
df$id <- as.numeric(df$id)

#turn off scientific notation
options(scipen = 999)

#make difference score
df$diffEC.L_MD <- df$EC.L_MD.02 - df$EC.L_MD.01

#restructure data for RCT completers only (N=41) #Note: we decided NOT to remove the CT outlier (participant 210030)
RCT_MD_EC <- df[(df$category == "RCT"),]

#restructure data for RCT & Relapse participants (N=74)
RCTRelapse_MD_EC <- df %>%
  gather(oldcolname, metric_t, ends_with(".01"), ends_with(".02")) %>%
  separate(oldcolname, into = c("metric", "timepoint"), sep = "_") %>%
  mutate(model_days = if_else(str_sub(timepoint, -1, -1)=="1", 1, dateDiff)) 

#write out clean dataframe
write.csv(RCT_MD_EC, '../generated_data/df_MD_leftEC.csv', row.names=FALSE)

```

## RCT only

```{r RCT_lEC, warning = FALSE}

#boxplot of difference in MD in EC (y axis) by randomization group (x axis)
(plot <- ggplot(RCT_MD_EC, aes(x= randomization, y = diffEC.L_MD)) + 
   geom_boxplot(outlier.shape = NA) + geom_jitter() +
   ggtitle("Mean diffusivity in left entorhinal cortex") +
   xlab("randomization condition") +  
   ylab("difference between timepoints")) 

#Note: the outlying participant is again 210030; this participant has NOT been excluded from subsequent stats

#run linear model without covariates
fit_rct <- lm(diffEC.L_MD ~ randomization + (1|id), data= RCT_MD_EC)
print(fit_rct)
summary(fit_rct)

#run linear model with covariates of sex and age
fit_rct <- lm(diffEC.L_MD ~ randomization + sex + age + (1|id), data= RCT_MD_EC)
print(fit_rct)
summary(fit_rct)

```

## RCT & Relapse (with time as factor)

```{r RCTRelapse_lEC, warning = FALSE}

#plot
RCTRelapse_MD_EC %>%
  ggplot(aes(x=model_days, y=metric_t, colour=randomization)) + 
  geom_point() + 
  geom_line(aes(group=id), alpha = 0.5) + 
  geom_smooth(method="lm", formula=y~poly(x,1)) +
  ggtitle("Mean diffusivity in left entorhinal cortex over time") +
      xlab("days between MRIs") +  
      ylab("thickness value") +
  theme_minimal() 

#run mixed linear model, with covariates
fit_all <- lmer(metric_t ~ randomization*model_days + sex + age + (1|id), data= RCTRelapse_MD_EC)
summary(fit_all)

#cleanup
rm(df, fit_all, fit_rct, MD, MD_wide, plot, RCT_MD_EC, RCTRelapse_MD_EC, completers)

```
